Moving create rule responsibility from EditRuleController to AppliedRuleController
